// apps/cf-worker/cloud-storage/src/types/types.d.ts
export interface AuthConfig {
  prefixes: string[]
  actives: Array<{
    name: string
    displayName: string
  }>
}

export interface FileInfo {
  size: number
  name: string
  key: string
  mime: string
  displayName?: string
  createdAt: Number
  updatedAt: Number
}

export interface ApiResponse<T> {
  code: number
  size?: number
  data?: T
  error?: string
  message?: string
}

export interface ListFilesRequestBody {
  pwd: string
}

export type ListFilesResponseBody = ApiResponse<FileInfo[]> & { size: number }